audio: {Coverity}:android U /hardware/amlogic/LibAudio/amadec/ [1/1]
PD#SWPL-123658

Problem:
{Coverity}:android U /hardware/amlogic/LibAudio/amadec/

Solution:
1.Null pointer dereferences
2.Resource leaks
3.Concurrent data access violations
4.Uninitialized variables

Verify:
U ohm

+++ LibAudio/amadec/adec_write.h
@@ -28,7 +28,7 @@
 
 typedef struct buffer_stream_st {
     int buf_length;
-    int buf_level;
+    unsigned long long buf_level;
     unsigned char * data;
     unsigned char * rd_ptr;
     unsigned char * wr_ptr;
